Feature: Generation of more detailed error pages
This set of tools generates a more informative set of error pages to be
used with the default backend.

Pages are color coded as follows:

- 1xx: Grey background
- 2xx: Green background
- 3xx: Blue background
- 4xx: Yellow background
- 5xx: Red background

Beyond the color coding, informational messages originating from the RFC
documents (with the exception of the message for error code 418 - "I am
a teapot") and a link to the relevant section are embedded.