Rework MIME/extension handling and debugging
This re-tools the MIME/extension handling to search for all valid
extensions as dictated by the included MIME types file.  Furthermore it
better uses the "DEBUG" environment variable to limit spurious log
messages.

The process of crash-looping through the MIME extensions could be better
done, but this is "good enough for me".

Put differently... Patches are accepted. 😎