repo: Restructure pass two
It was always intended that another restructuring pass would need to
occur as this is pulling files from multiple repos (and I wanted to
maintain the commit history as much as possible out of respect to the
previous authors).

In this change the Go server's module path has been updated, and the
path "images" removed.  Additionally, the examples for Kubernetes have
been moved more shallowly into the docs path and it's been made
(slightly) more clear that "user-guide.md" is more accurately
information on how data will be fed in from the NGINX ingress controller
("info-nginx-ingress.md").