Project rename: Penalty Flag
If for no other reason than to make it easier to document the project is
being renamed to "Penalty Flag".

The idea is that, in the same way that the color of a referee's flag bears
meaning for particpants and spectators of a football game, this custom error
server provides more colorful (and generally useful) "flags" when something goes
wrong with a request processed by ingress.